[types] Migrate function arg parser from cli types to rooch-types
